feat: move default container helper out of PodMenuItem
